+++ inst/shiny/DiagnosticsExplorer/server.R
@@ -80,17 +80,17 @@ shiny::shinyServer(function(input, output, session) {
   })
   
   shiny::observe({
-    subset <- unique(conceptSets$conceptSetName[conceptSets$cohortId == cohortId()])
+    subset <- unique(conceptSets$conceptSetName[conceptSets$cohortId == cohortId()]) %>% sort()
     shinyWidgets::updatePickerInput(session = session,
                                     inputId = "conceptSet",
                                     choices = subset)
   })
   
-  output$phenoTypeDescriptionTable <- DT::renderDataTable({
+  output$phenoTypeDescriptionTable <- DT::renderDataTable(expr = {
     data <- phenotypeDescription %>% 
-      dplyr::mutate(literatureReview = dplyr::case_when(.data$literatureReview != '' ~ 
+      dplyr::mutate(literatureReview = dplyr::case_when(!.data$literatureReview %in% c('','0') ~ 
                                                           paste0("<a href='", .data$literatureReview, "' target='_blank'>", "Link", "</a>"),
-                                                        TRUE ~ '')) %>%
+                                                        TRUE ~ 'Ongoing')) %>%
       dplyr::mutate(referentConceptId = paste0("<a href='", paste0(conceptIdBaseUrl(), .data$referentConceptId), "' target='_blank'>", .data$referentConceptId, "</a>")) 
     
     options = list(pageLength = 20,
@@ -108,15 +108,16 @@ shiny::shinyServer(function(input, output, session) {
                                filter = c("bottom"),
                                class = "stripe compact")
     return(dataTable)
-  })
+  }, server = TRUE)
   
-  output$cohortDescriptionTable <- DT::renderDataTable({
+  output$cohortDescriptionTable <- DT::renderDataTable(expr = {
     data <- cohortDescription %>% 
       dplyr::mutate(atlasId = as.integer(.data$atlasId)) %>% #this is temporary - we need to standardize this 
       dplyr::left_join(y = phenotypeDescription) %>% 
       dplyr::left_join(y = cohort, by = c('atlasId' = 'cohortId')) %>% #this is temporary - we need to standardize this 
       dplyr::mutate(cohortFullName = paste0("<a href='", paste0(cohortBaseUrl(), .data$atlasId),"' target='_blank'>", paste0(.data$cohortDefinitionName), "</a>")) %>% 
-      dplyr::select(phenotypeId, phenotypeName, cohortDefinitionId, cohortFullName, logicDescription, cohortDefinitionNotes)
+      dplyr::select(phenotypeId, phenotypeName, cohortDefinitionId, cohortFullName, logicDescription, cohortDefinitionNotes) %>% 
+      dplyr::arrange(phenotypeId, phenotypeName, cohortDefinitionId, cohortFullName)
     
     options = list(pageLength = 20,
                    searching = TRUE,
@@ -133,14 +134,14 @@ shiny::shinyServer(function(input, output, session) {
                                filter = c("bottom"),
                                class = "stripe compact")
     return(dataTable)
-  })
+  }, server = TRUE)
   
-  output$cohortCountsTable <- DT::renderDataTable({
+  output$cohortCountsTable <- DT::renderDataTable(expr = {
     data <- cohortCount[cohortCount$databaseId %in% input$databases, ]
     if (nrow(data) == 0) {
       return(NULL)
     }
-    databaseIds <- unique(data$databaseId)
+    databaseIds <- unique(data$databaseId) %>% sort()
     table <- data[data$databaseId == databaseIds[1], c("cohortId", "cohortEntries", "cohortSubjects")]
     colnames(table)[2:3] <- paste(colnames(table)[2:3], databaseIds[1], sep = "_")
     if (length(databaseIds) > 1) {
@@ -152,10 +153,12 @@ shiny::shinyServer(function(input, output, session) {
     }
     table <- merge(cohort, table, all.x = TRUE)
     table$url <- paste0(cohortBaseUrl2(), table$cohortId)
-    table$cohortFullName <- paste0("<a href='", table$url, "' target='_blank'>", table$cohortFullName, " (", table$cohortId, ")", "</a>")
+    table$cohortFullName <- paste0("<a href='", table$url, "' target='_blank'>", table$cohortFullName, "</a>")
     table$cohortId <- NULL
     table$cohortName <- NULL
     table$url <- NULL
+    table <- table %>% 
+            dplyr::arrange(.data$cohortFullName)
     
     sketch <- htmltools::withTags(table(
       class = 'display',
@@ -202,7 +205,7 @@ shiny::shinyServer(function(input, output, session) {
                                    backgroundPosition = "center")
     }
     return(dataTable)
-  })
+  }, server = TRUE)
   
   filteredIncidenceRates <- shiny::reactive({
     data <- incidenceRate[incidenceRate$cohortId == cohortId() & 
@@ -262,8 +265,7 @@ shiny::shinyServer(function(input, output, session) {
     return(plot)
   })
   
-  output$incidenceRatePlot <- shiny::renderPlot({
-    
+  output$incidenceRatePlot <- shiny::renderPlot(expr = {
     return(incidentRatePlotDownload())
   }, res = 100)
   
@@ -341,16 +343,15 @@ shiny::shinyServer(function(input, output, session) {
     
   })
   
-  output$timeDisPlot <- shiny::renderPlot({
+  output$timeDisPlot <- shiny::renderPlot(expr = {
     return(timeDisPlotDownload())
-    
   }, res = 100)
   
-  output$timeDistTable <- DT::renderDataTable({
+  output$timeDistTable <- DT::renderDataTable(expr = {
     data <- timeDistribution[timeDistribution$cohortId == cohortId() & 
                                timeDistribution$databaseId %in% input$databases, ]
     if (nrow(data) == 0) {
-      return(NULL)
+      return(tidyr::tibble(' ' = paste0('No data available for selected databases and cohorts')))
     }
     columns <- c("timeMetric", "averageValue", "standardDeviation", "minValue", "p10Value", "p25Value", "medianValue", "p75Value", "p90Value", "maxValue")
     headers <- c("Time Measure", "Average", "SD", "Min", "P10", "P25", "Median", "P75", "P90", "Max")
@@ -376,12 +377,15 @@ shiny::shinyServer(function(input, output, session) {
     table <- DT::formatRound(table, c("averageValue", "standardDeviation"), digits = 2)
     table <- DT::formatRound(table, c("minValue", "p10Value", "p25Value", "medianValue", "p75Value", "p90Value", "maxValue"), digits = 0)
     return(table)
-  })
+  }, server = TRUE)
   
-  output$includedConceptsTable <- DT::renderDataTable({
+  output$includedConceptsTable <- DT::renderDataTable(expr = {
     table <- includedSourceConcept[includedSourceConcept$cohortId == cohortId() &
                                      includedSourceConcept$conceptSetName == input$conceptSet & 
                                      includedSourceConcept$databaseId == input$database, ]
+    if (nrow(table) == 0) {
+      return(tidyr::tibble(' ' = paste0('No data available for selected databases and cohorts')))
+    }
     if (input$includedType == "Source Concepts") {
       table <- table[, c("conceptSubjects", "sourceConceptId", "sourceVocabularyId", "conceptCode", "sourceConceptName")]
       table <- table[!is.na(table$sourceConceptName), ]
@@ -416,14 +420,14 @@ shiny::shinyServer(function(input, output, session) {
                              backgroundRepeat = "no-repeat",
                              backgroundPosition = "center")
     return(table)
-  })
+  }, server = TRUE)
   
-  output$orphanConceptsTable <- DT::renderDataTable({
+  output$orphanConceptsTable <- DT::renderDataTable(expr = {
     table <- orphanConcept[orphanConcept$cohortId == cohortId() &
                              orphanConcept$conceptSetName == input$conceptSet & 
                              orphanConcept$databaseId == input$database, ]
     if (nrow(table) == 0) {
-      return(NULL)
+      return(tidyr::tibble(' ' = paste0('No data available for selected databases and cohorts')))
     }
     table <- table[, c("conceptCount", "conceptId", "standardConcept", "vocabularyId", "conceptCode", "conceptName")]
     table <- table[order(-table$conceptCount), ]
@@ -450,16 +454,16 @@ shiny::shinyServer(function(input, output, session) {
                              backgroundRepeat = "no-repeat",
                              backgroundPosition = "center")
     return(table)
-  })
+  }, server = TRUE)
   
-  output$inclusionRuleTable <- DT::renderDataTable({
+  output$inclusionRuleTable <- DT::renderDataTable(expr = {
     table <- inclusionRuleStats[inclusionRuleStats$cohortId == cohortId() & inclusionRuleStats$databaseId == input$database, ]
-    if (nrow(table) == 0) {
-      return(NULL)
-    }
     table <- table[order(table$ruleSequenceId), ]
     table$cohortId <- NULL
     table$databaseId <- NULL
+    if (nrow(table) == 0) {
+      return(tidyr::tibble(' ' = paste0('No data available for selected databases and cohorts')))
+    }
     lims <- c(0, max(table$remainSubjects))
     table <- table[, c("ruleSequenceId", "ruleName", "meetSubjects", "gainSubjects", "totalSubjects", "remainSubjects")]
     colnames(table) <- c("Sequence", "Name", "Meet", "Gain", "Total", "Remain")
@@ -484,13 +488,13 @@ shiny::shinyServer(function(input, output, session) {
                              backgroundRepeat = "no-repeat",
                              backgroundPosition = "center")
     return(table)
-  })
+  }, server = TRUE)
   
-  output$breakdownTable <- DT::renderDataTable({
+  output$breakdownTable <- DT::renderDataTable(expr = {
     data <- indexEventBreakdown[indexEventBreakdown$cohortId == cohortId() & 
                                   indexEventBreakdown$databaseId %in% input$databases, ]
     if (nrow(data) == 0) {
-      return(NULL)
+      return(tidyr::tibble(' ' = paste0('No data available for selected databases and cohorts')))
     }
     data <- data[, c("conceptId", "conceptName", "conceptCount", "databaseId" )]
     databaseIds <- unique(data$databaseId)
@@ -530,9 +534,9 @@ shiny::shinyServer(function(input, output, session) {
                                    backgroundPosition = "center")
     }
     return(dataTable)
-  })
+  }, server = TRUE)
   
-  output$characterizationTable <- DT::renderDataTable({
+  output$characterizationTable <- DT::renderDataTable(expr = {
     data <- covariateValue %>% 
       dplyr::filter(.data$cohortId == cohortId() & 
                       .data$databaseId %in% input$databases) %>% 
@@ -545,7 +549,9 @@ shiny::shinyServer(function(input, output, session) {
                               dplyr::filter(.data$cohortId == cohortId()) %>% 
                               dplyr::select(-cohortId))) %>% 
       dplyr::arrange(.data$databaseId)
-    
+    if (nrow(dataCounts) == 0) {
+      return(tidyr::tibble(' ' = paste0('No data available for selected databases and cohorts')))
+    }
     if (input$charType == "Pretty") {
       data <- data %>% 
         dplyr::left_join(y = covariate) %>% 
@@ -608,7 +614,7 @@ shiny::shinyServer(function(input, output, session) {
                                backgroundRepeat = "no-repeat",
                                backgroundPosition = "center")
     } else {
-      table <- data %>% 
+      data <- data %>% 
         dplyr::mutate(databaseId = stringr::str_replace_all(string = .data$databaseId, pattern = "_", replacement = " ")) %>% 
         tidyr::pivot_wider(id_cols = 'covariateId', 
                            names_from = "databaseId",
@@ -616,10 +622,11 @@ shiny::shinyServer(function(input, output, session) {
                            names_sep = "_",
                            values_fill = 0
         ) %>%  
-        dplyr::left_join(y = covariate %>% dplyr::select(.data$covariateId, .data$covariateName, .data$conceptId)) %>%
+        dplyr::left_join(y = covariate %>% dplyr::select(.data$covariateId, .data$covariateName, .data$conceptId) %>% dplyr::distinct()) %>%
         dplyr::select(-covariateId) %>% 
         dplyr::relocate("covariateName", "conceptId") %>% 
-        dplyr::arrange(.data$covariateName) 
+        dplyr::arrange(.data$covariateName) %>% 
+        dplyr::distinct()
       
       options = list(pageLength = 20,
                      searching = TRUE,
@@ -633,23 +640,23 @@ shiny::shinyServer(function(input, output, session) {
                        minCellPercentDef(1:(length(dataCounts$databaseId)) + 1)
                      )
       )
-      sketch <- htmltools::withTags(table(
-       class = 'display',
-       thead(
-         tr(
-           th(rowspan = 2, 'Covariate Name'),
-           th(rowspan = 2, 'Concept Id'),
-           lapply(dataCounts$databaseId, th, colspan = 2, class = "dt-center")
-         ),
-         tr(
-           lapply(rep(c("Proportion"), nrow(dataCounts)), th)
-         )
-       )
-      ))
-      table <- DT::datatable(table,
+      # sketch <- htmltools::withTags(table(
+      #  class = 'display',
+      #  thead(
+      #    tr(
+      #      th(rowspan = 2, 'Covariate Name'),
+      #      th(rowspan = 2, 'Concept Id'),
+      #      lapply(dataCounts$databaseId, th, colspan = 2, class = "dt-center")
+      #    ),
+      #    tr(
+      #      lapply(rep(c("Proportion"), nrow(dataCounts)), th)
+      #    )
+      #  )
+      # ))
+      table <- DT::datatable(data,
                              options = options,
                              rownames = FALSE,
-                             container = sketch, 
+                          #   container = sketch, 
                              escape = FALSE,
                              filter = c('bottom'),
                              class = "stripe nowrap compact")
@@ -661,9 +668,9 @@ shiny::shinyServer(function(input, output, session) {
                                backgroundPosition = "center")
     }
     return(table)
-  })
+  }, server = TRUE)
   
-  output$temporalCharacterizationTable <- DT::renderDataTable({
+  output$temporalCharacterizationTable <- DT::renderDataTable(expr = {
     table <- temporalCovariateValue %>% 
       dplyr::filter(.data$cohortId == cohortId(),
                     .data$databaseId == input$database,
@@ -671,7 +678,7 @@ shiny::shinyServer(function(input, output, session) {
       dplyr::select(-cohortId) %>% 
       dplyr::mutate(databaseId = stringr::str_replace_all(string = .data$databaseId, pattern = "_", replacement = " ")) %>% 
       dplyr::left_join(y = temporalCovariateChoices) %>% 
-      dplyr::arrange(timeId) %>%
+      dplyr::arrange(.data$timeId) %>%
       tidyr::pivot_wider(id_cols = 'covariateId', 
                          names_from = "choices",
                          values_from = "mean" ,
@@ -682,6 +689,10 @@ shiny::shinyServer(function(input, output, session) {
       dplyr::select(-.data$covariateId) %>% 
       dplyr::relocate(.data$covariateName, .data$conceptId) %>% 
       dplyr::arrange(.data$covariateName)
+    
+    if (nrow(table) == 0) {
+      return(tidyr::tibble(' ' = paste0('No data available for selected databases and cohorts')))
+    }
     
     temporalCovariateChoicesSelected <- temporalCovariateChoices %>% 
                                 dplyr::filter(.data$timeId %in% c(timeId())) 
@@ -698,23 +709,23 @@ shiny::shinyServer(function(input, output, session) {
                      minCellPercentDef(1:(length(temporalCovariateChoicesSelected$choices)) + 1)
                    )
     )
-    sketch <- htmltools::withTags(table(
-      class = 'display',
-      thead(
-        tr(
-          th(rowspan = 2, 'Covariate Name'),
-          th(rowspan = 2, 'Concept Id'),
-          lapply(temporalCovariateChoicesSelected$choices, th, colspan = 2, class = "dt-center")
-        ),
-        tr(
-          lapply(rep(c("Proportion"), length(temporalCovariateChoicesSelected$choices)), th)
-        )
-      )
-    ))
+    # sketch <- htmltools::withTags(table(
+    #   class = 'display',
+    #   thead(
+    #     tr(
+    #       th(rowspan = 2, 'Covariate Name'),
+    #       th(rowspan = 2, 'Concept Id'),
+    #       lapply(temporalCovariateChoicesSelected$choices, th, colspan = 2, class = "dt-center")
+    #     ),
+    #     tr(
+    #       lapply(rep(c("Proportion"), length(temporalCovariateChoicesSelected$choices)), th)
+    #     )
+    #   )
+    # ))
     table <- DT::datatable(table,
                            options = options,
                            rownames = FALSE,
-                           container = sketch,
+                          # container = sketch,
                            escape = FALSE,
                            filter = c('bottom'),
                            class = "stripe nowrap compact")
@@ -725,9 +736,9 @@ shiny::shinyServer(function(input, output, session) {
                              backgroundRepeat = "no-repeat",
                              backgroundPosition = "center")
     return(table)
-  })
+  }, server = TRUE)
   
-  output$overlapTable <- DT::renderDataTable({
+  output$overlapTable <- DT::renderDataTable(expr = {
     data <- cohortOverlap[cohortOverlap$targetCohortId == cohortId() & 
                             cohortOverlap$comparatorCohortId == comparatorCohortId() &
                             cohortOverlap$databaseId == input$database, ]
@@ -771,7 +782,7 @@ shiny::shinyServer(function(input, output, session) {
                            filter = c('bottom'),
                            class = "stripe nowrap compact")
     return(table)
-  })
+  }, server = TRUE)
   
   overLapPlot <- shiny::reactive({
     data <- cohortOverlap[cohortOverlap$targetCohortId == cohortId() & 
@@ -801,7 +812,7 @@ shiny::shinyServer(function(input, output, session) {
     return(plot)
   })
   
-  output$overlapPlot <- shiny::renderPlot({
+  output$overlapPlot <- shiny::renderPlot(expr = {
     return(overLapPlot())
   }, res = 100)
   
@@ -822,7 +833,7 @@ shiny::shinyServer(function(input, output, session) {
     return(balance)
   })
   
-  output$charCompareTable <- DT::renderDataTable({
+  output$charCompareTable <- DT::renderDataTable(expr = {
     balance <- computeBalance()
     if (nrow(balance) == 0) {
       return(NULL)
@@ -830,7 +841,7 @@ shiny::shinyServer(function(input, output, session) {
     
     if (input$charCompareType == "Pretty table") {
       table <- prepareTable1Comp(balance) %>% 
-        dplyr::arrange(sortOrder) %>% 
+        dplyr::arrange(.data$sortOrder) %>% 
         dplyr::select(-.data$sortOrder)
       
       options = list(pageLength = 999,
@@ -864,10 +875,10 @@ shiny::shinyServer(function(input, output, session) {
     } else {
       table <- balance %>% 
         dplyr::arrange(.data$covariateName) %>% 
-        dplyr::select(.data$covariateName, .data$conceptId, .data$mean1, .data$sd1, .data$mean2, .data$sd2, .data$stdDiff) %>% 
+        dplyr::select(.data$covariateName, .data$mean1, .data$mean2, .data$stdDiff, .data$conceptId) %>% 
         dplyr::mutate(stdDiff = round(x = .data$stdDiff, digits = 3)) %>% 
-        dplyr::rename_with(.fn = ~ stringr::str_replace(string = ., pattern = '1', replacement = 'Target')) %>% 
-        dplyr::rename_with(.fn = ~ stringr::str_replace(string = ., pattern = '2', replacement = 'Comparator')) %>% 
+        dplyr::rename_with(.fn = ~ stringr::str_replace(string = ., pattern = 'mean1', replacement = 'Target')) %>% 
+        dplyr::rename_with(.fn = ~ stringr::str_replace(string = ., pattern = 'mean2', replacement = 'Comparator')) %>% 
         dplyr::rename_with(.fn = SqlRender::camelCaseToTitleCase)
       
       options = list(pageLength = 20,
@@ -879,7 +890,7 @@ shiny::shinyServer(function(input, output, session) {
                      paging = TRUE,
                      columnDefs = list(
                        truncateStringDef(0, 150),
-                       minCellRealDef(c(1,3), 2)
+                       minCellPercentDef(1:2)
                      )
       )
       table <- DT::datatable(table,
@@ -889,21 +900,21 @@ shiny::shinyServer(function(input, output, session) {
                              filter = c('bottom'),
                              class = "stripe nowrap compact")
       table <- DT::formatStyle(table = table,
-                               columns = c(2,4),
+                               columns = 2:3,
                                background = DT::styleColorBar(c(0,1), "lightblue"),
                                backgroundSize = "98% 88%",
                                backgroundRepeat = "no-repeat",
                                backgroundPosition = "center")
       table <- DT::formatStyle(table = table,
-                               columns = 6,
+                               columns = 4,
                                background = styleAbsColorBar(1, "lightblue", "pink"),
                                backgroundSize = "98% 88%",
                                backgroundRepeat = "no-repeat",
                                backgroundPosition = "center")
-      table <- DT::formatRound(table, c(3, 5, 6), digits = 2)
+      table <- DT::formatRound(table, 4, digits = 2)
     }
     return(table)
-  })
+  }, server = TRUE)
   
   downloadCohortComparePlot <- shiny::reactive({
     balance <- computeBalance()
@@ -923,7 +934,7 @@ shiny::shinyServer(function(input, output, session) {
     return(plot)
   })
   
-  output$charComparePlot <- shiny::renderPlot({
+  output$charComparePlot <- shiny::renderPlot(expr = {
     return(downloadCohortComparePlot())
   }, res = 100)
   
@@ -976,7 +987,7 @@ shiny::shinyServer(function(input, output, session) {
   #   return(text)
   # })
   
-  output$databaseInformationTable <- DT::renderDataTable({
+  output$databaseInformationTable <- DT::renderDataTable(expr = {
     
     table <- database[, c("databaseId", "databaseName", "description")]
     options = list(pageLength = 20,
@@ -993,7 +1004,7 @@ shiny::shinyServer(function(input, output, session) {
                            rownames = FALSE,
                            class = "stripe compact")
     return(table)
-  })
+  }, server = TRUE)
   
   showInfoBox <- function(title, htmlFileName) {
     shiny::showModal(shiny::modalDialog(
@@ -1066,7 +1077,7 @@ shiny::shinyServer(function(input, output, session) {
     
     return(htmltools::withTags(
       div(
-        h5("Target: ", targetCohortWithCount$cohortFullName, " ( n = ", targetCohortWithCount$cohortSubjects, " )")
+        h5("Target: ", targetCohortWithCount$cohortFullName, " ( n = ", scales::comma(x = targetCohortWithCount$cohortSubjects), " )")
       )
     )
     )
@@ -1087,11 +1098,11 @@ shiny::shinyServer(function(input, output, session) {
       div(table(
         tr(
           td(
-            h5("Target: ", targetCohortWithCount$cohortFullName, " ( n = ", targetCohortWithCount$cohortSubjects, " )"),
+            h5("Target: ", targetCohortWithCount$cohortFullName, " ( n = ", scales::comma(targetCohortWithCount$cohortSubjects), " )"),
           ),
           td(HTML("&nbsp;&nbsp;&nbsp;&nbsp;")),
           td(
-            h5("Comparator : ", comparatorCohortWithCount$cohortFullName, " ( n = ", comparatorCohortWithCount$cohortSubjects, " )")
+            h5("Comparator : ", comparatorCohortWithCount$cohortFullName, " ( n = ",scales::comma(comparatorCohortWithCount$cohortSubjects), " )")
           )
         )
       )